Conflicts:
	src/main/webapp/WEB-INF/spring-web.xml

+++ src/main/java/evdc/vianet/ticket/service/TicketCommentService.java
@@ -2,10 +2,10 @@ package evdc.vianet.ticket.service;
 
 import java.util.List;
 
-import evdc.vianet.ticket.entity.TicketMessageView;
+import evdc.vianet.ticket.entity.view.TicketMessageView;
 
 public interface TicketCommentService {
-	public int addTicketComment(long ticketId, long userId, long teamId, String message, String scope);
+	public long addTicketComment(long ticketId, long userId, long teamId, String message, String scope);
 	public List<TicketMessageView> getCommentsByTicketIdAndScope(long ticketId, String scope);
 	public List<TicketMessageView> getCommentsByTicketIdAndteamIdAndScope(long ticketId, long teamId, String scope);
 }
